requirements.txt
# Django и базовые
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-two-factor-auth==1.15.1

# Безопасность и шифрование
cryptography==42.0.5
pyotp==2.9.0
qrcode==7.4.2
django-ratelimit==4.0.0
django-ipware==5.0.0

# Мониторинг и логирование
sentry-sdk==1.40.0
geoip2==4.8.0
psutil==5.9.6

# Утилиты
python-dotenv==1.0.0
Pillow==10.1.0
celery==5.3.4python
redis==5.0.1

# Разработка
ipython==8.17.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3